controller/service 수정중

+++ moit-v3/back/src/main/java/com/moit/member/service/UserService.java
@@ -21,23 +21,23 @@ public interface UserService {
 	public List<UserDto> select10(Map<String,Object> paramMap); // 회원리스트 페이징 조회
 	public int selectCnt(Map<String,Object> paramMap); // 조건별 전체 회원 수 조회
 	public int deleteUser(String loginId); // 관리자기능 - 회원 삭제
-	public boolean deleteMember(int memberId , String password);
+	public boolean deleteMember(Long memberId , String password);
 	public UserDto findByLoginId(UserDto dto);// 마이페이지
 	public UserDto findId(UserDto dto); // 아이디 찾기
 	public UserDto findPasswordUser(UserDto dto); // 비밀번호 찾기(정보조회)
 	public boolean changePassword(UserDto dto); // 비밀번호 재발급(변경)
 	//boolean changePassword( int memberId, String currentPassword, String newPassword ); // 비밀번호 변경
-	PasswordChangeResult changePassword(int memberId, String currentPassword, String newPassword);
-	public UserDto findByMemberId( int memberId);
+	PasswordChangeResult changePassword(Long memberId, String currentPassword, String newPassword);
+	public UserDto findByMemberId( Long memberId);
 	
 	/* security login */
 	public AuthUserDto readByLoginId(UserDto dto); //로그인	
 	public void insertSocialInfo(UserDto dto);
 	
 	// 추후 삭제될 수도 있는 코드
-	public List<String> getInterestList(int memberId);//관심사
+	public List<String> getInterestList(Long memberId);//관심사
 	//public void completeSocialJoin(UserDto dto);
 	List<InterestDto> getAllInterest();
-	List<Integer> getInterestIds(Integer memberId);
-	public void updateInterest( Integer memberId, List<Integer> interestIds );
+	List<Integer> getInterestIds(Long memberId);
+	public void updateInterest( Long memberId, List<Integer> interestIds );
 }
